backend/requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv
bcrypt
pyjwt